requirements.txt
fastapi
uvicorn
xmltodict
google-generativeai